requirements.txt
# Core dependencies for AI Trilingual Coach
google-genai>=0.8.0
notion-client>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0